# swagger/swagger.json
{
  "swagger": "2.0" ,
  "info": {
    "version": "1.0.0" ,
    "title": "Buddy App" ,
    "description": "Buddy Swagger" ,
    "license": {
      "name": "MIT" ,
      "url": "https://opensource.org/licenses/MIT"
    }
  } ,
  "host": "192.168.200.34:3000" ,
...
  "schemes": [
    "http"
  ] ,
  "consumes": [
    "application/json"
  ] ,
  "produces": [
    "application/json"
  ] ,
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security":[
    {
      "api_key":[]
    }
  ],
  "paths": {
    "/user/userLogin":{
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Login route for users",
        "description": "Allows user to login by providing email and password",
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "description": "login credentials of user",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses" :{
          "201": {
            "description": "User logged in successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/teacher":{
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Add teacher",
        "description": "Add new teacher to system",
        "parameters": [
          {
            "name": "insert teacher",
            "in": "body",
            "description": "teacher data",
            "schema": {
              "$ref": "#/definitions/teacherData"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses" :{
          "201": {
            "description": "Teacher created successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "get":{
        "tags":[
          "User"
        ],
        "summary":"Get teachers",
        "description": "Get all teachers",
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Teachers fetched",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/parentStudent":{
      "post":{
        "tags":[
          "User"
        ],
        "summary": "Create parent and student",
        "description": "Add new parent and student to the system",
        "parameters":[
          {
            "name": "create student and teacher",
            "in": "body",
            "description": "parent and student data",
            "schema":{
              "$ref": "#/definitions/parentStudentData"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Student and teacher added to system successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/assignTeacher":{
      "post":{
        "tags":[
          "User"
        ],
        "summary": "Assign class to teacher",
        "description": "Make teacher class teacher by assigning class",
        "parameters":[
          {
            "name": "Assign teacher",
            "in": "body",
            "description": "teacher and class info",
            "schema": {
              "$ref": "#/definitions/assignClassData"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "201":{
            "description":"Teacher assigned successfully",
            "schema":{
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/unassignTeacher/{teacher_id}":{
      "delete":{
        "tags":[
          "User"
        ],
        "summary": "Un-Assign class to teacher",
        "description": "Un-assign teacher as class teacher",
        "parameters":[
          {
            "name": "teacher_id",
            "in": "path",
            "description": "teacher id to be unassigned",
            "required": true,
            "type" : "integer"
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "201":{
            "description":"Teacher assigned successfully",
            "schema":{
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/changePassword":{
      "put":{
        "tags": [
          "User"
        ],
        "summary": "Change password",
        "description": "Allows the user of the system to change password",
        "parameters":[
            {
              "name": "password",
              "in": "body",
              "description": "new password",
              "schema": {
                "$ref": "#/definitions/passwordChange"
              }
            }
          ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Password changed",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/deleteUser":{
      "delete":{
        "tags": [
          "User"
        ],
        "summary": "Delete user",
        "description": "Delete user(student and parent) from system",
        "parameters":[
          {
            "name": "emails",
            "in": "body",
            "description": "emails of users to be removed",
            "schema":{
              "$ref": "#/definitions/removeUsers"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Users removed changed",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/updateStudent":{
      "post":{
        "tags":[
          "User"
        ],
        "summary": "Update student",
        "description": "Update student detail",
        "parameters": [
          {
            "name": "Update data",
            "in": "body",
            "description": "New data of user",
            "schema": {
              "$ref": "#/definitions/updateUserData"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Student updated",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/user/forgotPassword/{email}": {
      "get":{
        "tags": [
          "User"
        ],
        "summary": "Forgot password",
        "description":"Get code for setting new password",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Email of user",
            "type": "string"
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Code sent on user email",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/resetPassword": {
      "put":{
        "tags": [
          "User"
        ],
        "summary": "Set new password",
        "description": "Allows the user to set new password when user forgets password",
        "parameters":[
          {
            "name": "password",
            "in": "body",
            "description": "new password",
            "schema": {
              "$ref": "#/definitions/resetPassword"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses":{
          "200": {
            "description": "Password changed",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
  } ,
  "definitions": {
    "UserLogin": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "pattern": "^(([^<>()\\[\\]\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "teacherData":{
        "required": [
          "email",
          "name",
          "mobile_no",
          "address"
        ],
      "properties":{
        "email": {
          "type": "string",
          "pattern": "^(([^<>()\\[\\]\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$"
        },
        "name": {
          "type": "string"
        },
        "mobile_no": {
          "type": "string"
        },
        "address": {
          "type" : "string"
        }
      }
    },
    "parentStudentData":{
      "required": [
        "parent_email",
        "student_email",
        "parent_name",
        "phone_number",
        "student_name",
        "standard_id",
        "division_id"
      ],
      "properties":{
        "parent_email":{
          "type": "string",
          "pattern": "^(([^<>()\\[\\]\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$"
        },
        "student_email":{
          "type": "string",
          "pattern": "^(([^<>()\\[\\]\\.,;:\\s@\"]+(\\.[^<>()\\[\\]\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$"
        },
        "parent_name": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "student_name":{
          "type": "string"
        },
        "standard_id":{
          "type": "integer"
        },
        "division_id":{
          "type": "integer"
        }
      }
    },
    "assignClassData":{
      "required":[
        "teacher_id",
        "standard_id",
        "division_id"
      ],
      "properties": {
        "teacher_id":{
          "type": "integer"
        },
        "standard_id":{
          "type": "integer"
        },
        "division_id": {
          "type":"integer"
        }
      }
    },
    "passwordChange":{
      "required":[
        "oldPassword",
        "password"
      ],
      "properties":{
        "oldPassword": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "removeUsers":{
      "required":[
        "student_email",
        "parent_email"
      ],
      "properties": {
        "student_email":{
          "type": "string"
        },
        "parent_email": {
          "type": "string"
        }
      }
    },
    "updateUserData":{
      "required":[
        "studentName",
        "parentName",
        "phone_number",
        "studentEmail",
        "parentEmail"
      ],
      "properties":{
        "studentName": {
          "type": "string"
        },
        "parentName": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "studentEmail": {
          "type": "string"
        },
        "parentEmail": {
          "type": "string"
        }
      }
    },
...
    "resetPassword" :{
      "required": [
        "email",
        "password"
      ],
      "properties" :{
        "email" :{
          "type": "string"
        },
        "password" :{
          "type": "string"
        }
      }
    }
  }
}